requirements.txt
pvporcupine
pyaudio
git+https://github.com/openai/whisper.git
openai
requests
python-dotenv 